refactor(stack): stack sheet to support multiple
