+++ src/components/stacks/Sheet/index.tsx
@@ -1,31 +1,12 @@
-import { Codesnippet } from "@/components/CodeSnippet";
-import { CollapsibleCard } from "@/components/CollapsibleCard";
-import { SheetHeader } from "@/components/sheet/SheetHeader";
 import { ResizableSheetContent } from "@/components/sheet/resizable-sheet";
-import { stackQueries } from "@/data/stacks";
-import { extractComponents } from "@/lib/components";
-import { snakeCaseToTitleCase } from "@/lib/strings";
-import { sanitizeUrl } from "@/lib/url";
-import { routes } from "@/router/routes";
-import { StackComponent } from "@/types/components";
-import { useQuery } from "@tanstack/react-query";
-import {
-	Avatar,
-	AvatarFallback,
-	Box,
-	Button,
-	Sheet,
-	SheetTrigger,
-	Skeleton
-} from "@zenml-io/react-component-library";
-import { PropsWithChildren, useEffect, useState } from "react";
-import { Link } from "react-router-dom";
-import { CopyButton } from "../../CopyButton";
-import { Numberbox } from "../../NumberBox";
-import { ComponentBadge } from "../../stack-components/ComponentBadge";
-import { IntegrationsContextProvider, useIntegrationsContext } from "./IntegrationsContext";
-import { UpdateButtonContent } from "@/components/buttons/update-button-content";
-import { NestedCollapsible } from "@/components/NestedCollapsible";
+import { SheetHeader } from "@/components/sheet/SheetHeader";
+import { Sheet, SheetTrigger } from "@zenml-io/react-component-library";
+import { PropsWithChildren, useState } from "react";
+import { ComponentList } from "./ComponentList";
+import { IntegrationsContextProvider } from "./IntegrationsContext";
+import { StackHeadline } from "./StackHeadline";
+import { StackLabels } from "./StackLabels";
+import { StackSetCommand } from "./StackSetCommand";
 
 type Props = {
 	stackId: string;
@@ -50,196 +31,5 @@ export function StackSheet({
 				</IntegrationsContextProvider>
 			</ResizableSheetContent>
 		</Sheet>
-	);
-}
-
-function StackHeadline({ stackId }: Props) {
-	const stack = useQuery({ ...stackQueries.stackDetail(stackId) });
-
-	if (stack.isError) return null;
-	if (stack.isPending)
-		return (
-			<div className="p-5">
-				<Skeleton className="h-6 w-full" />
-			</div>
-		);
-
-	return (
-		<div className="flex min-w-[10rem] items-center justify-between border-b border-theme-border-moderate bg-theme-surface-primary p-5">
-			<div className="flex items-center space-x-2">
-				<Avatar type="square" size="lg" className="shrink-0">
-					<AvatarFallback size="lg">{stack.data.name[0]}</AvatarFallback>
-				</Avatar>
-				<div>
-					<div className="group/copybutton grid grid-cols-1 items-center gap-0.5">
-						<p className="mb-0.5 truncate text-text-sm text-theme-text-secondary">
-							{stack.data.id}
-						</p>
-						<CopyButton copyText={stack.data.id} />
-					</div>
-
-					<div className="flex items-center gap-1">
-						<h2 className="text-display-xs font-semibold">{stack.data.name}</h2>
-					</div>
-				</div>
-			</div>
-			<div className="flex items-center gap-2">
-				<Button asChild type="button" size="sm" intent="primary" emphasis="subtle">
-					<Link to={routes.stacks.edit(stackId)}>
-						<UpdateButtonContent />
-					</Link>
-				</Button>
-			</div>
-		</div>
-	);
-}
-
-function ComponentList({ stackId }: Props) {
-	const { setIntegrations } = useIntegrationsContext();
-	const stack = useQuery({ ...stackQueries.stackDetail(stackId) });
-
-	useEffect(() => {
-		if (!stack.data) return;
-
-		const components = extractComponents(
-			stack.data.metadata?.components as Record<string, StackComponent[]> | undefined
-		);
-
-		const integrations = components
-			.map((component) => component.body?.integration)
-			.filter(
-				(integration): integration is string =>
-					!!integration && integration !== "built-in" && integration !== "custom"
-			);
-
-		if (integrations.length >= 1) {
-			setIntegrations((prev) => Array.from(new Set([...prev, ...integrations])));
-		}
-	}, [stack.data, setIntegrations]);
-
-	if (stack.isError) return null;
-	if (stack.isPending)
-		return (
-			<div className="p-5">
-				<Skeleton className="h-[300px] w-full" />
-			</div>
-		);
-
-	const components = extractComponents(
-		stack.data.metadata?.components as Record<string, StackComponent[]> | undefined
-	);
-
-	return (
-		<ul className="space-y-5 p-5">
-			{components.map((component) => (
-				<li key={component.id}>
-					<ComponentListItem component={component} />
-				</li>
-			))}
-		</ul>
-	);
-}
-
-type ComponentListItemProps = {
-	component: StackComponent;
-};
-function ComponentListItem({ component }: ComponentListItemProps) {
-	return (
-		<Box className="flex items-center justify-between p-5">
-			<div className="flex items-center space-x-3">
-				<img
-					width={32}
-					height={32}
-					alt={`${component.resources?.flavor?.body?.display_name} logo`}
-					src={sanitizeUrl(component.body?.logo_url || "")}
-				/>
-				<div>
-					<Link to={routes.components.detail(component.id)} className="text-text-xl">
-						{component.name}
-					</Link>
-					<div className="group/copybutton flex items-center gap-0.5">
-						<p className="text-text-sm text-theme-text-secondary">{component.id.split("-")[0]}</p>
-						<CopyButton copyText={component.id} />
-					</div>
-				</div>
-			</div>
-			<Link to={routes.components.overview}>
-				<ComponentBadge type={component.body?.type || "orchestrator"}>
-					{snakeCaseToTitleCase(component.body?.type || "")}
-				</ComponentBadge>
-			</Link>
-		</Box>
-	);
-}
-
-type StackSetCommandProps = {
-	name: string;
-};
-function StackSetCommand({ name }: StackSetCommandProps) {
-	const { integrations } = useIntegrationsContext();
-
-	return (
-		<section className="px-5 pt-5">
-			<CollapsibleCard title="Set this stack" initialOpen>
-				<ul className="space-y-5">
-					<li className="space-y-2">
-						<div className="flex items-center gap-2">
-							<Numberbox>1</Numberbox>
-							<p className="font-semibold">Set your stack</p>
-						</div>
-						<div className="space-y-1">
-							<p className="text-text-sm text-theme-text-secondary">
-								Set the stack as active on your client
-							</p>
-							<Codesnippet
-								codeClasses="whitespace-pre-wrap"
-								wrap
-								code={`zenml stack set ${name}`}
-							/>
-						</div>
-					</li>
-					{integrations.length >= 1 && (
-						<li className="space-y-2">
-							<div className="flex items-center gap-2">
-								<Numberbox>2</Numberbox>
-								<p className="font-semibold">Install the integrations</p>
-							</div>
-							<div className="space-y-1">
-								<p className="text-text-sm text-theme-text-secondary">
-									Install the required integrations to run pipelines in your stack
-								</p>
-								<Codesnippet
-									codeClasses="whitespace-pre-wrap"
-									wrap
-									code={`zenml integration install ${integrations.join(" ")} --uv`}
-								/>
-							</div>
-						</li>
-					)}
-				</ul>
-			</CollapsibleCard>
-		</section>
-	);
-}
-
-function StackLabels({ stackId }: Props) {
-	const stack = useQuery({ ...stackQueries.stackDetail(stackId) });
-
-	if (stack.isError) return null;
-	if (stack.isPending)
-		return (
-			<section className="px-5">
-				<Skeleton className="h-[150px] w-full" />
-			</section>
-		);
-
-	const labels = stack.data.metadata?.labels;
-
-	if (!labels || Object.keys(labels).length === 0) return null;
-
-	return (
-		<section className="px-5">
-			<NestedCollapsible intent="primary" isInitialOpen title="Labels" data={labels} />
-		</section>
 	);
 }
